💥 Return document as inserted in database
Migration Guide:

If you previously relied on having the document passed to `insert` ammended with the `_id` added, and then having that same object returned, you need to rewrite your code to not depend on that behaviour. From now on, the objects passed in are not modified in any way.

The returned documents from `insert` now matches what is saved in the datbase, instead of being the same instances as passed in. In practice this means that e.g. classes will be converted into plain objects.